Use tar version.txt instead of yaml defined version.txt
This solves an issue in which FEDC is unable to update the version.txt declaration in YAML.
To solve this, this commit reapplies apply_extra, which task is to now extract the version.txt from deep in the tar.
This will fix the files being extracted over and over again, slowing launch times.